requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
matplotlib==3.7.2
plotly==5.17.0
scikit-image==0.21.0
